pyproject.toml
[project]
name = "avs-toolkit"
version = "0.1.0"
description = "A toolkit for orchestrating Agentic Value Streams."
authors = [{ name = "Patrick Heaney" }]
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "rich>=13.7.0",
    "marko>=2.0.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "mcp>=0.1.0",
    "python-dotenv>=1.2.1",
    "pypandoc>=1.16.2",
    "python-docx>=1.2.0",
    "pyperclip>=1.11.0",
]
requires-python = ">=3.12"

[project.scripts]
avs = "avs_toolkit.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
